Moving data model object and attribute definition classes to data model package.
