+++ dao/src/main/java/com/hashmapinc/server/dao/datamodel/AttributeDefinitionService.java
@@ -13,11 +13,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hashmapinc.server.dao.service.datamodelobject.nosql;
+package com.hashmapinc.server.dao.datamodel;
 
-import com.hashmapinc.server.dao.service.DaoNoSqlTest;
-import com.hashmapinc.server.dao.service.datamodelobject.DataModelObjectServiceImpTest;
+import com.hashmapinc.server.common.data.DataModelObject.AttributeDefinition;
+import com.hashmapinc.server.common.data.id.DataModelObjectId;
 
-@DaoNoSqlTest
-public class DataModelObjectServiceNoSqlTest extends DataModelObjectServiceImpTest {
+import java.util.List;
+import java.util.UUID;
+
+public interface AttributeDefinitionService {
+    AttributeDefinition save(AttributeDefinition attributeDefinition);
+    AttributeDefinition findById(UUID id);
+    List<AttributeDefinition> findByDataModelObjectId(DataModelObjectId dataModelObjectId);
+    boolean deleteById(UUID id);
 }
