+++ dao/src/main/java/com/hashmapinc/server/dao/datamodel/AttributeDefinitionServiceImp.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hashmapinc.server.dao.datamodelobject;
+package com.hashmapinc.server.dao.datamodel;
 
 import com.hashmapinc.server.common.data.DataModelObject.AttributeDefinition;
 import com.hashmapinc.server.common.data.DataModelObject.DataModelObject;
