+++ dao/src/main/java/com/hashmapinc/server/dao/datamodel/CassandraBaseAttributeDefinitionDao.java
@@ -13,43 +13,45 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hashmapinc.server.dao.sql.DataModelObject;
+package com.hashmapinc.server.dao.datamodel;
 
+import com.datastax.driver.core.querybuilder.Select;
 import com.hashmapinc.server.common.data.DataModelObject.AttributeDefinition;
-import com.hashmapinc.server.common.data.UUIDConverter;
 import com.hashmapinc.server.common.data.id.DataModelObjectId;
 import com.hashmapinc.server.dao.DaoUtil;
-import com.hashmapinc.server.dao.datamodelobject.AttributeDefinitionDao;
-import com.hashmapinc.server.dao.model.sql.AttributeDefinitionEntity;
-import com.hashmapinc.server.dao.sql.JpaAbstractDao;
-import com.hashmapinc.server.dao.util.SqlDao;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.repository.CrudRepository;
+import com.hashmapinc.server.dao.model.ModelConstants;
+import com.hashmapinc.server.dao.model.nosql.AttributeDefinitionEntity;
+import com.hashmapinc.server.dao.nosql.CassandraAbstractModelDao;
+import com.hashmapinc.server.dao.util.NoSqlDao;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.UUID;
 
-@Service
-@SqlDao
-public class JpaAttributeDefinitionDao extends JpaAbstractDao<AttributeDefinitionEntity, AttributeDefinition> implements AttributeDefinitionDao{
+import static com.datastax.driver.core.querybuilder.QueryBuilder.eq;
+import static com.datastax.driver.core.querybuilder.QueryBuilder.select;
 
-    @Autowired
-    AttributeDefinitionRepository attributeDefinitionRepository;
+@Service
+@NoSqlDao
+public class CassandraBaseAttributeDefinitionDao extends CassandraAbstractModelDao<AttributeDefinitionEntity, AttributeDefinition> implements AttributeDefinitionDao{
 
     @Override
-    protected Class<AttributeDefinitionEntity> getEntityClass() {
+    protected Class getColumnFamilyClass() {
         return AttributeDefinitionEntity.class;
     }
 
     @Override
-    protected CrudRepository<AttributeDefinitionEntity, String> getCrudRepository() {
-        return attributeDefinitionRepository;
+    protected String getColumnFamilyName() {
+        return ModelConstants.ATTRIBUTE_DEFINITION_COLUMN_FAMILY_NAME;
     }
+
 
     @Override
     public List<AttributeDefinition> findByDataModelObjectId(DataModelObjectId dataModelObjectId) {
-        List<AttributeDefinitionEntity> entities = attributeDefinitionRepository.findByDataModelObjectId(UUIDConverter.fromTimeUUID(dataModelObjectId.getId()));
+        Select select = select().from(ModelConstants.ATTRIBUTE_DEFINITION_COLUMN_FAMILY_NAME).allowFiltering();
+        Select.Where query = select.where();
+        query.and(eq(ModelConstants.ATTRIBUTE_DEFINITION_MODEL_OBJECT_ID, dataModelObjectId.getId()));
+        List<AttributeDefinitionEntity> entities = findListByStatement(query);
         return DaoUtil.convertDataList(entities);
     }
