+++ dao/src/main/java/com/hashmapinc/server/dao/datamodel/CassandraBaseDataModelObjectDao.java
@@ -13,50 +13,52 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hashmapinc.server.dao.datamodelobject;
+package com.hashmapinc.server.dao.datamodel;
 
 import com.datastax.driver.core.querybuilder.Select;
-import com.hashmapinc.server.common.data.DataModelObject.AttributeDefinition;
+import com.hashmapinc.server.common.data.DataModelObject.DataModelObject;
 import com.hashmapinc.server.common.data.id.DataModelObjectId;
+import com.hashmapinc.server.common.data.id.TenantId;
 import com.hashmapinc.server.dao.DaoUtil;
 import com.hashmapinc.server.dao.model.ModelConstants;
-import com.hashmapinc.server.dao.model.nosql.AttributeDefinitionEntity;
-import com.hashmapinc.server.dao.nosql.CassandraAbstractModelDao;
+import com.hashmapinc.server.dao.model.nosql.DataModelObjectEntity;
+import com.hashmapinc.server.dao.nosql.CassandraAbstractSearchTextDao;
 import com.hashmapinc.server.dao.util.NoSqlDao;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.UUID;
 
 import static com.datastax.driver.core.querybuilder.QueryBuilder.eq;
 import static com.datastax.driver.core.querybuilder.QueryBuilder.select;
 
 @Service
+@Slf4j
 @NoSqlDao
-public class CassandraBaseAttributeDefinitionDao extends CassandraAbstractModelDao<AttributeDefinitionEntity, AttributeDefinition> implements AttributeDefinitionDao{
+public class CassandraBaseDataModelObjectDao extends CassandraAbstractSearchTextDao<DataModelObjectEntity, DataModelObject> implements DataModelObjectDao {
 
     @Override
-    protected Class getColumnFamilyClass() {
-        return AttributeDefinitionEntity.class;
+    protected Class<DataModelObjectEntity> getColumnFamilyClass() {
+        return DataModelObjectEntity.class;
     }
 
     @Override
     protected String getColumnFamilyName() {
-        return ModelConstants.ATTRIBUTE_DEFINITION_COLUMN_FAMILY_NAME;
+        return ModelConstants.DATA_MODEL_OBJECT_CF;
     }
 
+    @Override
+    public DataModelObject findById(DataModelObjectId id) {
+        return super.findById(id.getId());
+    }
 
     @Override
-    public List<AttributeDefinition> findByDataModelObjectId(DataModelObjectId dataModelObjectId) {
-        Select select = select().from(ModelConstants.ATTRIBUTE_DEFINITION_COLUMN_FAMILY_NAME).allowFiltering();
+    public List<DataModelObject> findByTenantId(TenantId tenantId) {
+        Select select = select().from(ModelConstants.DATA_MODEL_OBJECT_CF).allowFiltering();
         Select.Where query = select.where();
-        query.and(eq(ModelConstants.ATTRIBUTE_DEFINITION_MODEL_OBJECT_ID, dataModelObjectId.getId()));
-        List<AttributeDefinitionEntity> entities = findListByStatement(query);
+        query.and(eq(ModelConstants.DATA_MODEL_OBJECT_TENANT_ID_PROPERTY, tenantId.getId()));
+        List<DataModelObjectEntity> entities = findListByStatement(query);
         return DaoUtil.convertDataList(entities);
     }
 
-    @Override
-    public boolean deleteById(UUID id) {
-        return super.removeById(id);
-    }
 }
