+++ dao/src/main/java/com/hashmapinc/server/dao/datamodel/DataModelObjectService.java
@@ -13,18 +13,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hashmapinc.server.dao.datamodelobject;
+package com.hashmapinc.server.dao.datamodel;
 
 import com.hashmapinc.server.common.data.DataModelObject.DataModelObject;
 import com.hashmapinc.server.common.data.id.DataModelObjectId;
 import com.hashmapinc.server.common.data.id.TenantId;
 
 import java.util.List;
-import java.util.UUID;
 
-public interface DataModelObjectDao {
-    DataModelObject findById(DataModelObjectId id);
+public interface DataModelObjectService {
     DataModelObject save(DataModelObject dataModelObject);
+    DataModelObject findById(DataModelObjectId dataModelObjectId);
     List<DataModelObject> findByTenantId(TenantId tenantId);
-    boolean removeById(UUID id);
+    boolean deleteById(DataModelObjectId dataModelObjectId);
 }
