+++ dao/src/main/java/com/hashmapinc/server/dao/sql/datamodel/AttributeDefinitionRepository.java
@@ -13,14 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hashmapinc.server.dao.sql.DataModelObject;
+package com.hashmapinc.server.dao.sql.datamodel;
 
-import com.hashmapinc.server.dao.model.sql.DataModelObjectEntity;
+import com.hashmapinc.server.dao.model.sql.AttributeDefinitionEntity;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface DataModelObjectRespository extends CrudRepository<DataModelObjectEntity, String> {
-    List<DataModelObjectEntity> findByTenantId(@Param("tenantId") String tenantId);
+public interface AttributeDefinitionRepository extends CrudRepository<AttributeDefinitionEntity, String> {
+    List<AttributeDefinitionEntity> findByDataModelObjectId(@Param("dataModelObjectId") String dataModelObjectId);
 }
