+++ dao/src/test/java/com/hashmapinc/server/dao/service/AbstractServiceTest.java
@@ -28,7 +28,7 @@ import com.hashmapinc.server.common.data.plugin.ComponentDescriptor;
 import com.hashmapinc.server.dao.TagMetaData.TagMetaDataService;
 import com.hashmapinc.server.dao.cluster.NodeMetricService;
 import com.hashmapinc.server.dao.datamodel.DataModelService;
-import com.hashmapinc.server.dao.datamodelobject.DataModelObjectService;
+import com.hashmapinc.server.dao.datamodel.DataModelObjectService;
 import com.hashmapinc.server.dao.theme.ThemeService;
 import com.hashmapinc.server.dao.logo.LogoService;
 import com.hashmapinc.server.dao.user.UserService;
