+++ dao/src/test/java/com/hashmapinc/server/dao/service/datamodel/sql/DataModelObjectServiceSqlTest.java
@@ -13,17 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hashmapinc.server.dao.datamodelobject;
+package com.hashmapinc.server.dao.service.datamodel.sql;
 
-import com.hashmapinc.server.common.data.DataModelObject.AttributeDefinition;
-import com.hashmapinc.server.common.data.id.DataModelObjectId;
+import com.hashmapinc.server.dao.service.DaoSqlTest;
+import com.hashmapinc.server.dao.service.datamodel.DataModelObjectServiceImpTest;
 
-import java.util.List;
-import java.util.UUID;
-
-public interface AttributeDefinitionService {
-    AttributeDefinition save(AttributeDefinition attributeDefinition);
-    AttributeDefinition findById(UUID id);
-    List<AttributeDefinition> findByDataModelObjectId(DataModelObjectId dataModelObjectId);
-    boolean deleteById(UUID id);
+@DaoSqlTest
+public class DataModelObjectServiceSqlTest extends DataModelObjectServiceImpTest {
 }
